+++ src/prefect/agent/agent.py
@@ -461,7 +461,7 @@ class Agent:
             import asyncio
 
             try:
-                asyncio.get_event_loop()
+                asyncio.get_running_loop()
             except RuntimeError:
                 asyncio.set_event_loop(asyncio.new_event_loop())