+++ src/prefect/tasks/transform/__init__.py
@@ -1,0 +1,12 @@
+"""
+This module contains a collection of tasks to interact with Transform metrics layer.
+"""
+
+try:
+    from prefect.tasks.transform.transform_tasks import TransformCreateMaterialization
+except ImportError as err:
+    raise ImportError(
+        'Using `prefect.tasks.transform` requires Prefect to be installed with the "transform" extra.'
+    ) from err
+
+__all__ = ["TransformCreateMaterialization"]