+++ tests/engine/results/test_azure_result.py
@@ -17,6 +17,8 @@ class TestAzureResult:
         connection = MagicMock()
         azure = MagicMock(from_connection_string=connection)
         monkeypatch.setattr("azure.storage.blob.BlobServiceClient", azure)
+        credential = MagicMock(return_value="mocked!")
+        monkeypatch.setattr("azure.identity.DefaultAzureCredential", credential)
         yield connection
 
     def test_azure_init(self, azure_client):
@@ -32,15 +34,32 @@ class TestAzureResult:
         result = AzureResult(container="bob", connection_string="conn", value=3)
         assert result.value == 3
 
+    def test_azure_init_without_connection_string(self, azure_client):
+        result = AzureResult(container="bob", connection_string=None)
+        with pytest.raises(ValueError):
+            result.initialize_service()
+
     def test_azure_init_connection_string(self, azure_client):
+        result = AzureResult(container="bob", connection_string="con1;AccountKey=abc")
+        result.initialize_service()
+        azure_client.assert_called_with(conn_str="con1;AccountKey=abc", credential=None)
+
+        with prefect.context({"secrets": {"test": "con2;AccountKey=abc"}}):
+            result = AzureResult(container="bob", connection_string_secret="test")
+            result.initialize_service()
+            azure_client.assert_called_with(
+                conn_str="con2;AccountKey=abc", credential=None
+            )
+
+    def test_azure_init_connection_string_without_key(self, azure_client):
         result = AzureResult(container="bob", connection_string="con1")
         result.initialize_service()
-        azure_client.assert_called_with(conn_str="con1")
+        azure_client.assert_called_with(conn_str="con1", credential="mocked!")
 
         with prefect.context({"secrets": {"test": "con2"}}):
             result = AzureResult(container="bob", connection_string_secret="test")
             result.initialize_service()
-            azure_client.assert_called_with(conn_str="con2")
+            azure_client.assert_called_with(conn_str="con2", credential="mocked!")
 
     def test_azure_writes_to_blob_using_rendered_template_name(self, monkeypatch):
         client = MagicMock(upload_blob=MagicMock())